:construction: Variable Repository

+++ Spring_Data_JPA/src/main/java/fr/diginamic/Spring_Data_JPA/repository/AnimalRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import fr.diginamic.Spring_Data_JPA.model.Animal;
 
-public interface AnimalRepository extends CrudRepository<Animal, Long> {
+public interface AnimalRepository extends CrudRepository<Animal, Integer> {
 
 }
