+++ Spring_Data_JPA/src/main/java/fr/diginamic/Spring_Data_JPA/repository/SpeciesRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import fr.diginamic.Spring_Data_JPA.model.Species;
 
-public interface SpeciesRepository extends CrudRepository<Species, Long> {
+public interface SpeciesRepository extends CrudRepository<Species, Integer> {
 
 }
